+++ src/main/java/com/skolarli/lmsservice/services/impl/BatchScheduleServiceImpl.java
@@ -1,10 +1,13 @@
 package com.skolarli.lmsservice.services.impl;
 
+import com.skolarli.lmsservice.exception.OperationNotSupportedException;
 import com.skolarli.lmsservice.exception.ResourceNotFoundException;
-import com.skolarli.lmsservice.models.db.Batch;
 import com.skolarli.lmsservice.models.db.BatchSchedule;
+import com.skolarli.lmsservice.models.db.LmsUser;
 import com.skolarli.lmsservice.repository.BatchScheduleRepository;
 import com.skolarli.lmsservice.services.BatchScheduleService;
+import com.skolarli.lmsservice.utils.UserUtils;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -15,9 +18,11 @@ public class BatchScheduleServiceImpl implements BatchScheduleService {
     Logger logger = LoggerFactory.getLogger(BatchScheduleServiceImpl.class);
 
     final BatchScheduleRepository batchScheduleRepository;
+    final UserUtils userUtils;
 
-    public BatchScheduleServiceImpl(BatchScheduleRepository batchScheduleRepository) {
+    public BatchScheduleServiceImpl(BatchScheduleRepository batchScheduleRepository, UserUtils userUtils) {
         this.batchScheduleRepository = batchScheduleRepository;
+        this.userUtils = userUtils;
     }
 
     @Override
@@ -27,7 +32,18 @@ public class BatchScheduleServiceImpl implements BatchScheduleService {
 
     @Override
     public BatchSchedule updateBatchSchedule(BatchSchedule batchSchedule, long id) {
-        return null;
+        BatchSchedule existingBatchSchedule = batchScheduleRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("BatchSchedule", "Id", id));
+        LmsUser currentUser = userUtils.getCurrentUser();
+        if (currentUser.getIsAdmin() != true && currentUser != existingBatchSchedule.getBatch().getCourse().getCourseOwner()) {
+            throw new OperationNotSupportedException("Operation not supported: Permission Denied");
+        }
+        if (batchSchedule.getBatchScheduleIsDeleted() != null) {
+            logger.error("Cannot change deleted status. Use Delete API");
+            existingBatchSchedule.setBatchScheduleIsDeleted(null);
+        }
+        existingBatchSchedule.update(batchSchedule);
+        return batchScheduleRepository.save(existingBatchSchedule);
     }
 
     @Override
@@ -48,10 +64,33 @@ public class BatchScheduleServiceImpl implements BatchScheduleService {
         return batchSchedules;
     }
 
+    private Boolean checkPermissions (BatchSchedule existingBatchSchedule) {
+        LmsUser currentUser = userUtils.getCurrentUser();
+        if (currentUser.getIsAdmin() != true && currentUser != existingBatchSchedule.getBatch().getCourse().getCourseOwner()) {
+            return false;
+        }
+        return true;
+    }
+
     @Override
     public void deleteBatchSchedule(long id) {
         BatchSchedule existingBatchSchedule = batchScheduleRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("BatchSchedule", "Id", id));
+        if (checkPermissions(existingBatchSchedule) == false) {
+            throw new OperationNotSupportedException("Operation not supported: Permission Denied");
+        }
+        existingBatchSchedule.setBatchScheduleIsDeleted(true);
+        batchScheduleRepository.save(existingBatchSchedule);
+    }
+
+    @Override
+    public void hardDeleteBatchSchedule(long id) {
+        BatchSchedule existingBatchSchedule = batchScheduleRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("BatchSchedule", "Id", id));
+        if (checkPermissions(existingBatchSchedule) == false) {
+            throw new OperationNotSupportedException("Operation not supported: Permission Denied");
+        }
         batchScheduleRepository.delete(existingBatchSchedule);
+        
     }
 }
