+++ src/main/java/com/skolarli/lmsservice/services/impl/ChapterServiceImpl.java
@@ -5,23 +5,28 @@ import java.util.List;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import com.skolarli.lmsservice.exception.OperationNotSupportedException;
 import com.skolarli.lmsservice.exception.ResourceNotFoundException;
 import com.skolarli.lmsservice.models.db.Chapter;
+import com.skolarli.lmsservice.models.db.LmsUser;
 import com.skolarli.lmsservice.repository.ChapterRepository;
 import com.skolarli.lmsservice.services.ChapterService;
+import com.skolarli.lmsservice.utils.UserUtils;
 
 import ch.qos.logback.classic.Logger;
 
 @Service
 public class ChapterServiceImpl implements ChapterService {
 
-    Logger logger = (Logger) LoggerFactory.getLogger(ChapterServiceImpl.class);
+    private static final Logger logger = (Logger) LoggerFactory.getLogger(ChapterServiceImpl.class);
 
     ChapterRepository chapterRepository;
+    UserUtils userUtils;
 
-    public ChapterServiceImpl(ChapterRepository chapterRepository) {
+    public ChapterServiceImpl(ChapterRepository chapterRepository, UserUtils userutils) {
         super();
         this.chapterRepository = chapterRepository;
+        this.userUtils = userutils;
     }
 
     @Override
@@ -30,18 +35,47 @@ public class ChapterServiceImpl implements ChapterService {
         return chapterRepository.save(chapter);
     }
 
+    private Boolean checkPermission(Chapter chapter) {
+        LmsUser currentUser = userUtils.getCurrentUser();
+        if (currentUser.getIsAdmin() != true && currentUser != chapter.getCourse().getCourseOwner()) {
+            return false;
+        }
+        return true;
+    }
+
     @Override
-    public Chapter updateChapter(Chapter chapter) {
-        Chapter existingChapter = chapterRepository.findById(chapter.getId()).orElseThrow(
-                () -> new ResourceNotFoundException("Chapter", "Id", chapter.getId()));
+    public Chapter updateChapter(Chapter chapter, long id) {
+        Chapter existingChapter = chapterRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Chapter", "Id", id));
+        if (checkPermission(existingChapter) == false) {
+            throw new OperationNotSupportedException("User does not have permissions to update this chapter");
+        }
+        if (chapter.getChapterIsDeleted() != null) {
+            logger.error("Cannot change deleted status. Use delete APIs instead");
+            chapter.setChapterIsDeleted(null);
+        }
         existingChapter.update(chapter);
         return chapterRepository.save(existingChapter);
     }
 
     @Override
-    public void deleteChapter(Chapter chapter) {
-        Chapter existingChapter = chapterRepository.findById(chapter.getId()).orElseThrow(
-                () -> new ResourceNotFoundException("Chapter", "Id", chapter.getId()));
+    public void deleteChapter(long id) {
+        Chapter existingChapter = chapterRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Chapter", "Id", id));
+        if (checkPermission(existingChapter) == false) {
+            throw new OperationNotSupportedException("User does not have permissions to delete this chapter");
+        }
+        existingChapter.setChapterIsDeleted(true);
+        chapterRepository.save(existingChapter);
+    }
+
+    @Override
+    public void hardDeleteChapter(long id) {
+        Chapter existingChapter = chapterRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Chapter", "Id", id));
+        if (checkPermission(existingChapter) == false) {
+            throw new OperationNotSupportedException("User does not have permissions to delete this chapter");
+        }
         chapterRepository.delete(existingChapter);
     }
 
@@ -56,5 +90,4 @@ public class ChapterServiceImpl implements ChapterService {
     public List<Chapter> getAllChapters() {
         return chapterRepository.findAll();
     }
-    
 }
